requirements.txt
numpy
python-dotenv
pytest